Move AWS permissions functionality to the AWS provider
