+++ src/nimbo/core/cloud_provider/provider/services/storage.py
@@ -2,4 +2,7 @@ import abc
 
 
 class Storage(abc.ABC):
-    ...
+    @staticmethod
+    @abc.abstractmethod
+    def allow_ingress_current_ip(target: str, dry_run=False) -> None:
+        ...
