+++ src/nimbo/core/cloud_provider/provider_impl/gcp/services/gcp_permissions.py
@@ -5,3 +5,11 @@ class GcpPermissions(Permissions):
     @staticmethod
     def allow_ingress_current_ip(target: str, dry_run=False) -> None:
         ...
+
+    @staticmethod
+    def setup_as_user(role_name: str = "", dry_run=False) -> None:
+        pass
+
+    @staticmethod
+    def setup_as_admin(dry_run=False) -> None:
+        ...
