+++ src/nimbo/core/config/common.py
@@ -41,6 +41,7 @@ class BaseConfig(pydantic.BaseModel):
         title = "Nimbo configuration"
         extra = "forbid"
 
+    # TODO: this is not checked for
     provider: Optional[CloudProvider] = None
 
     local_datasets_path: Optional[str] = None
